Refactor developer view renderers into modules
