+++ src/ui/views/developer/dom.js
@@ -1,0 +1,8 @@
+export function setText(root, selector, value) {
+  const node = root?.querySelector?.(selector);
+  if (node) {
+    node.textContent = value;
+  }
+}
+
+export default setText;
